Refactoring : Moved BasicArrowButtonFixedSize as a private class of SwingSpinnerWidget
